More engineering function unit test refactorings
